add errorhandler namespace
