+++ api/oracle_framework.py
@@ -24,7 +24,7 @@ class OracleFramework:
 
     def has_results(self, full_response):
         """
-        Function used to determine if object returned from OracleFramework.get_prices() returns actual price data or empty dictionaries
+        Helper function used to determine if dictionary returned from OracleFramework.get_prices() contains price data or not
         """
         for source_response in full_response["sources"].values():
             if source_response: return True
